CG, delimiter и label в отдельную ЕТ
