+++ hohlova.polina/T2/delimiter.cpp
@@ -1,0 +1,36 @@
+#include "delimiter.hpp"
+#include <iostream>
+
+namespace hohlova
+{
+  std::istream& operator>>(std::istream& in, DelimiterIO&& dest)
+  {
+    std::istream::sentry sentry(in);
+    if (!sentry)
+    {
+      return in;
+    }
+    char c = '0';
+    in >> c;
+    if (in && (c != dest.exp))
+    {
+      in.setstate(std::ios::failbit);
+    }
+    return in;
+  }
+
+  std::istream& operator>>(std::istream& in, LabelIO&& dest)
+  {
+    std::istream::sentry sentry(in);
+    if (!sentry)
+    {
+      return in;
+    }
+    std::string data = "";
+    if ((in >> data) && (data != dest.exp))
+    {
+      in.setstate(std::ios::failbit);
+    }
+    return in;
+  }
+}
