+++ hohlova.polina/T2/structures.hpp
@@ -4,8 +4,6 @@
 #include <sstream>
 #include <string>
 #include <cassert>
-#include <iterator>
-#include <vector>
 #include <iomanip>
 
 namespace hohlova
@@ -19,11 +17,6 @@ namespace hohlova
     double key1;
     real key2;
     std::string key3;
-  };
-
-  struct DelimiterIO
-  {
-    char exp;
   };
 
   struct DoubleIO
@@ -41,16 +34,9 @@ namespace hohlova
     std::string& ref;
   };
 
-  struct LabelIO
-  {
-    std::string exp;
-  };
-
-  std::istream& operator>>(std::istream& in, DelimiterIO&& dest);
   std::istream& operator>>(std::istream& in, DoubleIO&& dest);
   std::istream& operator>>(std::istream& in, RealIO&& dest);
   std::istream& operator>>(std::istream& in, StringIO&& dest);
-  std::istream& operator>>(std::istream& in, LabelIO&& dest);
   std::istream& operator>>(std::istream& in, Data& dest);
   std::ostream& operator<<(std::ostream& out, const Data& dest);
